access control

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,10 @@ use App\Post;
 use DB;
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -48,6 +52,7 @@ class PostsController extends Controller
         $post=new Post;
         $post->title=$request->input('title');
         $post->body=$request->input('body');
+        $post->user_id= auth()->user()->id; 
         $post->save();
 
         return redirect('/posts')->with('success','Post Created Successfully');
@@ -75,6 +80,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post=Post::find($id);
+        if(auth()->user()->id==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -109,6 +117,9 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->id==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         $post=Post::find($id);
         $post->delete();
         return redirect('/posts')->with('success','Post deleted Successfully');
